+++ 3DBook/Models/FolderViewModel/BaseFolderViewModel.cs
@@ -4,7 +4,7 @@ public abstract class BaseFolderViewModel
 {
   
     public int Folds { get; set; }
-    public decimal Enter { get; set; }
-    public decimal Exit { get; set; }
+    public int Enter { get; set; }
+    public int Exit { get; set; }
 
 }
